refactor: cookie util 파일 이름 변경
