+++ src/components/pages/sign-in/Signin.jsx
@@ -8,7 +8,7 @@ import { toggleSnackbarOpen } from "@/redux/modules/snackbar";
 import Form from "@/components/common/form/Form";
 import Input from "@/components/common/input/Input";
 
-import { setCookie, getCookie } from "@/utils/auth";
+import { setCookie, getCookie } from "@/utils/cookie";
 import { BASE_URL, PATH, MESSAGE } from "@/constants";
 
 import StyledSigninContainer from "@/components/pages/sign-in/Signin.styled";
